requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.14
botocore==1.38.14
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dataclasses-json==0.6.7
datasets==2.21.0
dill==0.3.8
distro==1.9.0
dydantic==0.0.8
exceptiongroup==1.3.0
fastapi==0.115.12
fastmcp==2.3.3
filelock==3.18.0
Flask==3.1.0
fpdf==1.7.2
frozenlist==1.6.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
groq==0.24.0
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-anthropic==0.3.13
langchain-community==0.3.9
langchain-core==0.3.59
langchain-experimental==0.3.3
langchain-groq==0.3.2
langchain-mcp-adapters==0.0.10
langchain-neo4j==0.1.1
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langmem==0.0.25
langsmith==0.3.42
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.8.0
mcp-neo4j-cypher==0.2.1
mcp-neo4j-memory==0.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==1.38.2
neo4j==5.27.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
openai==1.78.1
openapi-pydantic==0.5.1
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
propcache==0.3.1
protobuf==4.25.7
pyarrow==20.0.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pypdf==4.3.1
python-a2a==0.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
s3transfer==0.12.0
safetensors==0.5.3
setuptools==80.3.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sse-starlette==2.3.4
starlette==0.46.2
streamlit==1.35.0
sympy==1.14.0
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
trustcall==0.0.39
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uv==0.7.3
uvicorn==0.34.2
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
